# Swagger Generated/Server/aspnetcore/src/IO.Swagger/Models/ErrorPayload.cs
/*
 * Ready4Tech API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text;
using Newtonsoft.Json;

namespace IO.Swagger.Models;

/// <summary>
/// </summary>
[DataContract]
public class ErrorPayload : IEquatable<ErrorPayload>
{
    /// <summary>
    ///     Gets or Sets Error
    /// </summary>
    [Required]
    [DataMember(Name = "error")]
    public ErrorPayloadError Error { get; set; }

    /// <summary>
    ///     Returns true if ErrorPayload instances are equal
    /// </summary>
    /// <param name="other">Instance of ErrorPayload to be compared</param>
    /// <returns>Boolean</returns>
    public bool Equals(ErrorPayload other)
    {
        if (ReferenceEquals(null, other)) return false;
        if (ReferenceEquals(this, other)) return true;

        return
            Error == other.Error ||
            (Error != null &&
             Error.Equals(other.Error));
    }

    /// <summary>
    ///     Returns the string presentation of the object
    /// </summary>
    /// <returns>String presentation of the object</returns>
    public override string ToString()
    {
        var sb = new StringBuilder();
        sb.Append("class ErrorPayload {\n");
        sb.Append("  Error: ").Append(Error).Append("\n");
        sb.Append("}\n");
        return sb.ToString();
    }

    /// <summary>
    ///     Returns the JSON string presentation of the object
    /// </summary>
    /// <returns>JSON string presentation of the object</returns>
    public string ToJson() => JsonConvert.SerializeObject(this, Formatting.Indented);

    /// <summary>
    ///     Returns true if objects are equal
    /// </summary>
    /// <param name="obj">Object to be compared</param>
    /// <returns>Boolean</returns>
    public override bool Equals(object obj)
    {
        if (ReferenceEquals(null, obj)) return false;
        if (ReferenceEquals(this, obj)) return true;
        return obj.GetType() == GetType() && Equals((ErrorPayload)obj);
    }

    /// <summary>
    ///     Gets the hash code
    /// </summary>
    /// <returns>Hash code</returns>
    public override int GetHashCode()
    {
        unchecked // Overflow is fine, just wrap
        {
            var hashCode = 41;
            // Suitable nullity checks etc, of course :)
            if (Error != null)
                hashCode = hashCode * 59 + Error.GetHashCode();
            return hashCode;
        }
    }

    #region Operators

#pragma warning disable 1591

    public static bool operator ==(ErrorPayload left, ErrorPayload right) => Equals(left, right);

    public static bool operator !=(ErrorPayload left, ErrorPayload right) => !Equals(left, right);

#pragma warning restore 1591

    #endregion Operators
}